old-tests/webapps/WebSockets/tests/approved/interfaces/WebSocket/url/003.html

<!doctype html>
<title>WebSockets: deleting url</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=../../../constants.js></script>
<div id=log></div>
<script>
test(function() {
  var ws = new WebSocket(SCHEME_AND_DOMAIN+'/');
  delete ws.url;
  assert_equals(ws.url, SCHEME_AND_DOMAIN+'/');
});
</script>
